// HW.Admin/ClientApp/src/app/Shared/Models/UserModel/adminuservm.ts
import { Data } from '@angular/router';
export class Adminuservm {
  id: string;
  userName: string;
  email: string;
  phoneNumber: number;
  createdOn: Date;
  role: string;
  lastActive: Data;
  password: string;
  securityRole: string;
  securityRoleId: string;
  securityRoleName: string;
  isTestUser: boolean;
  
}